classes refactor
